requirements.txt
Django~=5.0.6
python-dotenv~=1.0.1
django-crispy-forms~=2.1
cryptography~=42.0.7
PyMySQL~=1.1.1
mariadb~=1.1.10